ar(temp) auth stuff messy
