+++ lib/auth/constants.ts
@@ -6,15 +6,16 @@ import GithubProvider from 'next-auth/providers/github';
 import GoogleProvider from 'next-auth/providers/google';
 import AppleProvider from 'next-auth/providers/apple';
 import FacebookProvider from 'next-auth/providers/facebook';
-import EmailProvider from 'next-auth/providers/email';
+// import EmailProvider from 'next-auth/providers/email';
 import { PrismaAdapter } from '@auth/prisma-adapter';
+// import { authConfig } from './config';
 
-const providers = [
-  EmailProvider({
-    server: process.env.EMAIL_SERVER as string,
-    from: process.env.EMAIL_FROM as string,
-    // maxAge: 24 * 60 * 60, // How long email links are valid for (default 24h)
-  }),
+export const providers = [
+  // EmailProvider({
+  //   server: process.env.EMAIL_SERVER as string,
+  //   from: process.env.EMAIL_FROM as string,
+  //   // maxAge: 24 * 60 * 60, // How long email links are valid for (default 24h)
+  // }),
   GithubProvider({
     clientId: process.env.GITHUB_ID as string,
     clientSecret: process.env.GITHUB_SECRET as string,
@@ -33,21 +34,13 @@ const providers = [
   }),
 ];
 
-export const providerMap = providers.map((provider) => {
-  if (typeof provider === 'function') {
-    const providerData = provider();
-    return { id: providerData.id, name: providerData.name };
-  } else {
-    return { id: provider.id, name: provider.name };
-  }
-});
-
-export const authOptions = {
+// config.ts
+export const authConfig = {
   providers,
   adapter: PrismaAdapter(PrivatePrisma),
-  session: {
-    strategy: 'jwt',
-  },
+  // session: {
+  //   strategy: 'jwt',
+  // },
   events: {},
   callbacks: {
     async signIn() {
@@ -57,20 +50,20 @@ export const authOptions = {
     async redirect() {
       return `${process.env.MAIN_URL}`;
     },
-    async jwt({ user, token }) {
-      if (user) {
-        // Note that this if condition is needed
-        token.user = { ...user };
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token?.user) {
-        // Note that this if condition is needed
-        session.user = token.user;
-      }
-      return session;
-    },
+    // async jwt({ user, token }) {
+    //   if (user) {
+    //     // Note that this if condition is needed
+    //     token.user = { ...user };
+    //   }
+    //   return token;
+    // },
+    // async session({ session, token }) {
+    //   if (token?.user) {
+    //     // Note that this if condition is needed
+    //     session.user = token.user;
+    //   }
+    //   return session;
+    // },
   },
   cookies: {
     pkceCodeVerifier: {
@@ -92,4 +85,13 @@ export const authOptions = {
   },
 };
 
-export const { auth, handlers, signIn, signOut }: AuthOptions = NextAuth(authOptions);
+export const providerMap = providers.map((provider) => {
+  if (typeof provider === 'function') {
+    const providerData = provider();
+    return { id: providerData.id, name: providerData.name };
+  } else {
+    return { id: provider.id, name: provider.name };
+  }
+});
+
+export const { auth, handlers, signIn, signOut, config }: AuthOptions = NextAuth(authConfig);
